backend/requirements.txt
torch
transformers==4.56.2
datasets>=3.4.1,<4.0.0
huggingface_hub>=0.34.0
fastapi
uvicorn
python-multipart
librosa
soundfile
torchcodec
jinja2
langchain-openai
openai
pydantic
python-docx
python-dotenv
reportlab
sqlalchemy
psycopg2-binary
